feat: make mithras-crypto a lib
